refactor: centralize VM0 model credit multipliers
